Création de fonction offlineFilter

+++ coordinates.cpp
@@ -1,6 +1,6 @@
 #include "coordinates.h"
 
-Coordinates::Coordinates(int _x, int _y)
+Coordinates::Coordinates(double _x, double _y)
 	:col(_x), row(_y)
 {}
